Исправляю замечания: 	1) Исправил название файла локализации для русского языка 	2) Исправил название пакетов 	3) Логи на англ. языке 	4) public static final ... 	5) биндинг вью к презентеру вынесен в onViewCreated 	6) Очепятка NavigationDrawable -> Drawer 	7) Добавил unbinder в retained фрагменты.
