+++ app/build.gradle
@@ -41,16 +41,16 @@ dependencies {
     testCompile 'junit:junit:4.12'
 
     //butterknife
-    compile 'com.jakewharton:butterknife:8.2.1'
-    apt 'com.jakewharton:butterknife-compiler:8.2.1'
+    compile 'com.jakewharton:butterknife:8.7.0'
+    apt 'com.jakewharton:butterknife-compiler:8.7.0'
 
-    //navigation drawable
+    //navigation drawer
     compile('com.mikepenz:materialdrawer:5.9.4@aar') {
         transitive = true
     }
 
     //dagger 2
-    compile 'com.google.dagger:dagger:2.7'
-    apt 'com.google.dagger:dagger-compiler:2.7'
+    compile 'com.google.dagger:dagger:2.11'
+    apt 'com.google.dagger:dagger-compiler:2.11'
 
 }
