--- app/src/main/java/com/example/alexander/weatherapp/DI/Scopes/WeatherScope.java
@@ -1,15 +1,0 @@
-package com.example.alexander.weatherapp.DI.Scopes;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-import javax.inject.Scope;
-
-/**
- * Created by Alexander on 08.07.2017.
- */
-
-@Scope
-@Retention(RetentionPolicy.RUNTIME)
-public @interface WeatherScope {
-}
