--- app/src/main/java/com/example/alexander/weatherapp/Presentation/About/Interfaces/IAboutView.java
@@ -1,8 +1,0 @@
-package com.example.alexander.weatherapp.Presentation.About.Interfaces;
-
-/**
- * Created by Alexander on 08.07.2017.
- */
-
-public interface IAboutView {
-}
