--- app/src/main/java/com/example/alexander/weatherapp/Presentation/Weather/Interfaces/IWeatherView.java
@@ -1,10 +1,0 @@
-package com.example.alexander.weatherapp.Presentation.Weather.Interfaces;
-
-/**
- * Created by Alexander on 08.07.2017.
- */
-
-public interface IWeatherView {
-
-
-}
