+++ app/src/main/java/com/example/alexander/weatherapp/di/modules/AppModule.java
@@ -1,0 +1,30 @@
+package com.example.alexander.weatherapp.di.modules;
+
+import android.content.Context;
+import android.support.annotation.NonNull;
+
+import javax.inject.Singleton;
+
+import dagger.Module;
+import dagger.Provides;
+
+/**
+ * Created by Alexander on 08.07.2017.
+ */
+
+@Module
+public class AppModule {
+
+    private Context appContext;
+
+    public AppModule(@NonNull Context appContext) {
+        this.appContext = appContext;
+    }
+
+    @Provides
+    @Singleton
+    Context provideContext(){
+        return appContext;
+    }
+
+}
