+++ app/src/main/java/com/example/alexander/weatherapp/presentation/about/interfaces/IAboutPresenter.java
@@ -1,0 +1,12 @@
+package com.example.alexander.weatherapp.presentation.about.interfaces;
+
+/**
+ * Created by Alexander on 08.07.2017.
+ */
+
+public interface IAboutPresenter {
+
+    void bindView(IAboutView aboutView);
+    void unbindView();
+
+}
