+++ app/src/main/java/com/example/alexander/weatherapp/presentation/settings/intefaces/ISettingsPresenter.java
@@ -1,0 +1,12 @@
+package com.example.alexander.weatherapp.presentation.settings.intefaces;
+
+/**
+ * Created by Alexander on 08.07.2017.
+ */
+
+public interface ISettingsPresenter {
+
+    void bindView(ISettingsView settingsView);
+    void unbindView();
+
+}
