+++ app/src/main/java/com/example/alexander/weatherapp/presentation/settings/intefaces/ISettingsView.java
@@ -1,0 +1,8 @@
+package com.example.alexander.weatherapp.presentation.settings.intefaces;
+
+/**
+ * Created by Alexander on 08.07.2017.
+ */
+
+public interface ISettingsView {
+}
