+++ app/src/main/java/com/example/alexander/weatherapp/presentation/weather/interfaces/IWeatherView.java
@@ -1,0 +1,10 @@
+package com.example.alexander.weatherapp.presentation.weather.interfaces;
+
+/**
+ * Created by Alexander on 08.07.2017.
+ */
+
+public interface IWeatherView {
+
+
+}
