--- app/src/main/res/values-ru-rRU/strings.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <string name="weather">Погода</string>
-    <string name="settings">Настройки</string>
-    <string name="about">О программе</string>
-
-
-</resources>
